+++ tests/UnitTests/PhasorDynamics/runGovernorHygovTests.cpp
@@ -1,0 +1,17 @@
+#include "GovernorHygovTests.hpp"
+
+int main()
+{
+  GridKit::Testing::TestingResults result;
+
+  GridKit::Testing::GovernorHygovTests<double, size_t> test;
+
+  result += test.constructionAndValidation();
+  result += test.signals();
+  result += test.sourceDefault();
+  result += test.zeroTimeConstants();
+  result += test.baseConversion();
+  result += test.jsonParseAndSystemAssembly();
+
+  return result.summary();
+}